Changed InsertInvoiceSQL to accept a double. Removed create invoice button. Moved its functionality to save invoice. Removed the functionality of disabling controls when they shouldn't be used. There was a bug where the buttons remained grayed out even after being reenabled. I think this had something to do with the BureauBlack theme. Instead the functionality of these controls were removed if the requisiteconditions were not met. Moved the $ for prices to the label preciding the price instead of the label that holds the numerical price value to improve the ease of using said values. Bug fixes:    No longer crashes when save invoice is selected, but a date hasn't been selected or no items have been added.    No longer crashes when remove item button is clicked when no item was selected to be removed.    No longer crashes when add item button is clicked but no item was selected. TODO:    When Save invoice is clicked we need to distinguish between creating a new invoice and editing a selected invoice. When editing a selected invoice update it instead of creating a new one.    Fill in the date field when an invoice is selected.    There are a few unused functions and unnecessaryfunctions that can be removed.    XML documentation is still needed.

+++ Main/clsMainSQL.cs
@@ -50,7 +50,7 @@ namespace Invoices.Main
         /// SQL to insert into Invoices
         /// </summary>
         /// <returns>string containing SQL to insert values into Invoices</returns>
-        public string InsertInvoiceSQL(string date, int cost)
+        public string InsertInvoiceSQL(string date, double cost)
         {
             return "INSERT INTO Invoices (InvoiceDate, TotalCost) Values (#" + date + "#, " + cost + ")";
         }
